small update

+++ bot.js
@@ -88,7 +88,7 @@ bot.on(message("dice"), async (ctx) => {
     if(chat.id !== Number(idChat))return
 
     // without: || forward_from. for tests 
-    if(dice.emoji !== "🎲") return
+    if(dice.emoji !== "🎲" || forward_from) return
 
     if(Number(threadId) !== message_thread_id){
         await ctx.deleteMessage(message_id)
